added to users views
