+++ codesnippets_project/templates/base.html
@@ -5,12 +5,13 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Code Snippets</title>
+  <title>{% block title %}Code Snippets{% endblock %}</title>
   <link rel="stylesheet" href="{% static 'style.css' %}">
-
 </head>
 <body>
+  <main>
     {% block content %}
     {% endblock %}
+  </main>
 </body>
 </html>
