+++ codesnippets_project/templates/snippets.html
@@ -1,0 +1,1 @@
+{% extends "base.html" %}
