+++ codesnippets_project/users/admin.py
@@ -1,5 +1,10 @@
 from django.contrib import admin
+# from django.contrib.auth import get_user_model
+
+# from .forms import NewUserCreationForm, ChangeUserForm
 from .models import User
 # Register your models here.
+
+
 
 admin.site.register(User)
