+++ codesnippets_project/users/forms.py
@@ -1,9 +1,20 @@
-from django.contrib.auth.forms import UserCreationForm, UserChangeForm
+from django import forms
+from django.contrib.auth.forms import UserCreationForm, UserChangeForm, AuthenticationForm, PasswordResetForm, SetPasswordForm, PasswordChangeForm
 from .models import User
 
 class NewUserCreationForm(UserCreationForm):
-    pass
+    class Meta:
+        model = User
+        fields = [
+            'username',
+            'password',
+        ]
 
 class ChangeUserForm(UserChangeForm):
-    pass
+    class Meta:
+        model = User
+        fields = [
+            'username',
+            'password',
+        ]
