+++ codesnippets_project/users/models.py
@@ -4,3 +4,6 @@ from django.contrib.auth.models import AbstractUser
 
 class User(AbstractUser):
     pass
+
+    def __str__(self):
+        return f'{self.username}'
